Manage tweets with asterisks
There is an issue with sending tweet messages as form data when they contain an asterisk. See https://github.com/sferik/twitter/issues/677 for the equivalent issue in a ruby gem. This is because of a grey area between standards. We can resolve this by sending the tweet body as a URI encoded parameter instead of as form data in the body of the request.

This commit also cleans up some variable names to make it a bit clearer what things are doing, and added a couple of comments.

Fixes #2